Script Updates
-updated throw mechanic script
-merged levelselect.cs and mainmenubuttons.cs
-renamed mainmenubuttons to UIScript
-added FetchNames.cs (contains string constants)
