+++ Assets/Scripts/FetchScripts/FetchNames.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class FetchNames
+{
+    public const string BACKYARD_LEVEL = "BACKYARD_LEVEL";
+    public const string WAREHOUSE_LEVEL = "WAREHOUSE_LEVEL";
+    public const string ON_GAME_START = "ON_GAME_START";
+    public const string ON_RANDOM_START = "ON_RANDOM_START";
+}
