+++ Assets/Scripts/FetchScripts/FetchNames.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f238a977539d5e3459c3d4cc3d498638
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
