--- Assets/Scripts/FetchScripts/LevelSelect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ae1044dda91131341ba7325df57b82e3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
