--- Assets/Scripts/FetchScripts/MainMenuButtons.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c6c58b9e6e75f194786c90be43c5ee1f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
