+++ Assets/Scripts/FetchScripts/ThrowMechanic.cs
@@ -6,23 +6,15 @@ using UnityEngine;
 public class ThrowMechanic : MonoBehaviour
 {
     [SerializeField] private GameObject ball;
-    [SerializeField] private float groundLevel;
-    [SerializeField] private float throwX;
-    [SerializeField] private float throwY;
-    [SerializeField] private float throwZ;
+    [SerializeField] private Vector3 manualThrow;
+    [SerializeField] private Vector3 minRandomThrow;
+    [SerializeField] private Vector3 maxRandomThrow;
     private Rigidbody ballRigidbody;
-
-    [SerializeField] private float minX;
-    [SerializeField] private float maxX;
-    [SerializeField] private float minY;
-    [SerializeField] private float maxY;
-    [SerializeField] private float minZ;
-    [SerializeField] private float maxZ;
 
     private void Awake()
     {
-        EventBroadcaster.Instance.AddObserver(EventNames.FetchEvents.ON_GAME_START, this.throwBall);
-        EventBroadcaster.Instance.AddObserver(EventNames.FetchEvents.ON_RANDOM_START, this.randomThrow);
+        EventBroadcaster.Instance.AddObserver(FetchNames.ON_GAME_START, this.throwBall);
+        EventBroadcaster.Instance.AddObserver(FetchNames.ON_RANDOM_START, this.randomThrow);
         ballRigidbody = ball.GetComponent<Rigidbody>();
     }
 
@@ -37,23 +29,31 @@ public class ThrowMechanic : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-
+        EventBroadcaster.Instance.AddObserver(FetchNames.ON_GAME_START, this.throwBall);
+        EventBroadcaster.Instance.AddObserver(FetchNames.ON_RANDOM_START, this.randomThrow);
+        ballRigidbody = ball.GetComponent<Rigidbody>();
     }
 
     private void OnDestroy()
     {
-        EventBroadcaster.Instance.RemoveObserver(EventNames.FetchEvents.ON_GAME_START);
+        EventBroadcaster.Instance.RemoveObserver(FetchNames.ON_GAME_START);
+        EventBroadcaster.Instance.RemoveObserver(FetchNames.ON_RANDOM_START);
     }
 
     private void throwBall()
     {
-        ballRigidbody.AddForce(new Vector3(throwX, throwY, throwZ), ForceMode.Force);
+        ballRigidbody.AddForce(manualThrow, ForceMode.Force);
     }
 
     private void randomThrow()
     {
-        Vector3 randVector = new Vector3(Random.Range(minX, maxX), maxY, Random.Range(minZ, maxZ));
+        Vector3 randVector = new Vector3(Random.Range(
+                                            minRandomThrow.x, maxRandomThrow.x),
+                                          Random.Range(
+                                            minRandomThrow.y, maxRandomThrow.y),
+                                          Random.Range(
+                                              minRandomThrow.z, maxRandomThrow.z));
         ballRigidbody.AddForce(randVector);
     }
-    
+
 }
